Implement codegen for loading constant addresses

+++ src/main/java/net/rizecookey/racc0on/backend/x86_64/operand/x8664Immediate.java
@@ -1,7 +1,7 @@
 package net.rizecookey.racc0on.backend.x86_64.operand;
 
 public record x8664Immediate(String value) implements x8664Operand {
-    public x8664Immediate(int value) {
+    public x8664Immediate(long value) {
         this(String.valueOf(value));
     }
     @Override
